+++ l10n_br_nfe/__manifest__.py
@@ -27,6 +27,9 @@
         "views/res_company_view.xml",
         "views/nfe_document_view.xml",
         "views/res_config_settings_view.xml",
+        # Report
+        "report/reports.xml",
+        "report/danfe_nfce.xml",
     ],
     "demo": [
         "demo/res_users_demo.xml",
