+++ l10n_br_nfe/models/document.py
@@ -33,6 +33,7 @@ from odoo.addons.l10n_br_fiscal.constants.fiscal import (
     EVENT_ENV_HML,
     EVENT_ENV_PROD,
     EVENTO_RECEBIDO,
+    FISCAL_PAYMENT_MODE,
     LOTE_PROCESSADO,
     MODELO_FISCAL_NFCE,
     MODELO_FISCAL_NFE,
@@ -410,7 +411,14 @@ class NFe(spec_models.StackedModel):
     @api.depends("partner_id")
     def _compute_dest_data(self):
         for doc in self:  # TODO if out
-            doc.nfe40_dest = doc.partner_id
+            if (
+                doc.partner_id.is_anonymous_consumer
+                and not doc.partner_id.cnpj_cpf
+                and doc.document_type != MODELO_FISCAL_NFCE
+            ):
+                doc.nfe40_dest = None
+            else:
+                doc.nfe40_dest = doc.partner_id
 
     ##########################
     # NF-e tag: entrega
@@ -431,7 +439,13 @@ class NFe(spec_models.StackedModel):
     @api.depends("partner_shipping_id")
     def _compute_entrega_data(self):
         for rec in self:
-            rec.nfe40_entrega = rec.partner_shipping_id
+            if (
+                rec.document_type == MODELO_FISCAL_NFCE
+                and rec.partner_shipping_id.is_anonymous_consumer
+            ):
+                rec.nfe40_entrega = None
+            else:
+                rec.nfe40_entrega = rec.partner_shipping_id
 
     ##########################
     # NF-e tag: retirada
@@ -903,7 +917,26 @@ class NFe(spec_models.StackedModel):
                 _logger.error("DANFE Error \n {}".format(e))
         super()._exec_after_SITUACAO_EDOC_AUTORIZADA(old_state, new_state)
 
+    def _generate_key(self):
+        for record in self.filtered(filter_processador_edoc_nfe):
+            date = fields.Datetime.context_timestamp(record, record.document_date)
+            chave_edoc = ChaveEdoc(
+                ano_mes=date.strftime("%y%m").zfill(4),
+                cnpj_cpf_emitente=record.company_cnpj_cpf,
+                codigo_uf=(
+                    record.company_state_id and record.company_state_id.ibge_code or ""
+                ),
+                forma_emissao=int(self.nfe_transmission),
+                modelo_documento=record.document_type_id.code or "",
+                numero_documento=record.document_number or "",
+                numero_serie=record.document_serie or "",
+                validar=False,
+            )
+            record.document_key = chave_edoc.chave
+
     def _eletronic_document_send(self):
+        self._prepare_payments_for_nfce()
+
         super(NFe, self)._eletronic_document_send()
         for record in self.filtered(filter_processador_edoc_nfe):
             if record.xml_error_message:
@@ -956,6 +989,10 @@ class NFe(spec_models.StackedModel):
     def view_pdf(self):
         if not self.filtered(filter_processador_edoc_nfe):
             return super().view_pdf()
+
+        if self.document_type == MODELO_FISCAL_NFCE:
+            return self.action_danfe_nfce_report()
+
         if not self.authorization_file_id or not self.file_report_id:
             self.make_pdf()
         return self._target_new_tab(self.file_report_id)
@@ -1166,3 +1203,83 @@ class NFe(spec_models.StackedModel):
             return
 
         return self._processador().consulta_qrcode_url
+
+    def _prepare_payments_for_nfce(self):
+        for rec in self.filtered(lambda d: d.document_type == MODELO_FISCAL_NFCE):
+            rec.nfe40_detPag.filtered(lambda p: p.nfe40_tPag == "99").write(
+                {"nfe40_xPag": "Outros"}
+            )
+
+    def action_danfe_nfce_report(self):
+        return (
+            self.env["ir.actions.report"]
+            .search(
+                [("report_name", "=", "l10n_br_nfe.report_danfe_nfce")],
+                limit=1,
+            )
+            .report_action(self, data=self._prepare_nfce_danfe_values())
+        )
+
+    def _prepare_nfce_danfe_values(self):
+        return {
+            "company_ie": self.company_id.inscr_est,
+            "company_cnpj": self.company_id.cnpj_cpf,
+            "company_legal_name": self.company_id.legal_name,
+            "company_street": self.company_id.street,
+            "company_number": self.company_id.street_number,
+            "company_district": self.company_id.district,
+            "company_city": self.company_id.city_id.display_name,
+            "company_state": self.company_id.state_id.name,
+            "lines": self._prepare_nfce_danfe_line_values(),
+            "total_product_quantity": len(
+                self.fiscal_line_ids.filtered(lambda line: line.product_id)
+            ),
+            "amount_total": self.amount_total,
+            "amount_discount_value": self.amount_discount_value,
+            "amount_freight_value": self.amount_freight_value,
+            "payments": self._prepare_nfce_danfe_payment_values(),
+            "amount_change": self.nfe40_vTroco,
+            "nfce_url": self.get_nfce_qrcode_url(),
+            "document_key": self.document_key,
+            "document_number": self.document_number,
+            "document_serie": self.document_serie,
+            "document_date": self.document_date.astimezone().strftime(
+                "%d/%m/%y %H:%M:%S"
+            ),
+            "authorization_protocol": self.authorization_protocol,
+            "document_qrcode": self.get_nfce_qrcode(),
+            "system_env": self.nfe40_tpAmb,
+            "unformatted_amount_freight_value": self.amount_freight_value,
+            "unformatted_amount_discount_value": self.amount_discount_value,
+            "contingency": self.nfe_transmission != "1",
+            "homologation_environment": self.nfe_environment == "2",
+        }
+
+    def _prepare_nfce_danfe_line_values(self):
+        lines_list = []
+        lines = self.fiscal_line_ids.filtered(lambda line: line.product_id)
+        for index, line in enumerate(lines):
+            product_id = line.product_id
+            lines_list.append(
+                {
+                    "product_index": index + 1,
+                    "product_default_code": product_id.default_code,
+                    "product_name": product_id.name,
+                    "product_quantity": line.quantity,
+                    "product_uom": product_id.uom_name,
+                    "product_unit_value": product_id.lst_price,
+                    "product_unit_total": line.quantity * product_id.lst_price,
+                }
+            )
+        return lines_list
+
+    def _prepare_nfce_danfe_payment_values(self):
+        payments_list = []
+        for payment in self.nfe40_detPag:
+            payments_list.append(
+                {
+                    "method": dict(FISCAL_PAYMENT_MODE)[payment.nfe40_tPag],
+                    "value": payment.nfe40_vPag,
+                }
+            )
+        return payments_list
