+++ l10n_br_nfe/report/reports.xml
@@ -21,8 +21,8 @@
         <field name="name">DANFE NFC-e</field>
         <field name="model">account.move</field>
         <field name="report_type">qweb-html</field>
-        <field name="report_name">l10n_br_pos_nfce.report_danfe_nfce</field>
-        <field name="report_file">l10n_br_pos_nfce.danfe_nfce</field>
+        <field name="report_name">l10n_br_nfe.report_danfe_nfce</field>
+        <field name="report_file">l10n_br_nfe.danfe_nfce</field>
         <field name="paperformat_id" ref="danfe_nfce_qweb_paperformat" />
     </record>
