+++ l10n_br_pos_nfce/models/__init__.py
@@ -1,7 +1,3 @@
 from . import pos_config
 from . import pos_order
 from . import pos_payment_method
-from . import account_move
-from . import l10n_br_fiscal_document_workflow
-from . import res_partner
-from . import l10n_br_fiscal_document
