+++ l10n_br_pos_nfce/tests/__init__.py
@@ -1,5 +1,2 @@
 from . import common
 from . import test_pos_order
-from . import test_res_partner
-from . import test_fiscal_document
-from . import test_account_move
